refactor: move table query functions from _sessions to _queries module
