+++ .windsurf/workflows/check-mocks.md
@@ -1,0 +1,5 @@
+---
+description: Check that the right mocks are being used
+---
+
+For all test files, make sure that MagicMock and AsyncMock are used in the proper places.  Be careful to check async usage.
