+++ src/deephaven_mcp/sessions/__init__.py
@@ -27,7 +27,10 @@ from ._errors import SessionCreationError
 from ._sessions import (
     Session,
     SessionManager,
+)
+from ._queries import (
     get_dh_versions,
+    get_table,
     get_meta_table,
     get_pip_packages_table,
 )
@@ -35,6 +38,7 @@ from ._sessions import (
 __all__ = [
     "SessionManager",
     "get_dh_versions",
+    "get_table",
     "get_pip_packages_table",
     "get_meta_table",
     "SessionCreationError",
