+++ src/deephaven_mcp/sessions/_sessions.py
@@ -474,130 +474,3 @@ class SessionManager:
                 f"Session cached for worker: {worker_name}. Returning session."
             )
             return session
-
-
-async def get_table(session: Session, table_name: str) -> pyarrow.Table:
-    """
-    Retrieve a table from a Deephaven session as a pyarrow.Table.
-
-    Args:
-        session (Session): The Deephaven session to retrieve the table from.
-        table_name (str): The name of the table to retrieve.
-
-    Returns:
-        pyarrow.Table: The table as a pyarrow.Table.
-    """
-    table = await asyncio.to_thread(session.open_table, table_name)
-    return await asyncio.to_thread(table.to_arrow)
-
-
-async def get_meta_table(session: Session, table_name: str) -> pyarrow.Table:
-    """
-    Retrieve the meta table (schema/metadata) for a Deephaven table as a pyarrow.Table.
-
-    Args:
-        session (Session): The Deephaven session to retrieve the meta table from.
-        table_name (str): The name of the table to retrieve the meta table for.
-
-    Returns:
-        pyarrow.Table: The meta table containing schema/metadata information for the specified table.
-    """
-    table = await asyncio.to_thread(session.open_table, table_name)
-    # the lambda is needed to avoid the meta_table property from being evaluated outside the thread
-    meta_table = await asyncio.to_thread(lambda: table.meta_table)
-    return await asyncio.to_thread(meta_table.to_arrow)
-
-
-async def get_pip_packages_table(session: Session) -> pyarrow.Table:
-    """
-    Returns a table of installed pip packages from a Deephaven session.
-
-    Args:
-        session (Session):
-            An active Deephaven session in which to run the script and retrieve the resulting table.
-
-    Returns:
-        pyarrow.Table:
-            A pyarrow.Table containing two columns: 'Package' (str) and 'Version' (str), listing all installed pip packages.
-
-    Raises:
-        Exception: On failure to run the script or retrieve the table.
-
-    Example:
-        >>> arrow_table = await get_pip_packages_table(session)
-        >>> df = arrow_table.to_pandas()
-        >>> print(df.head())
-    """
-    script = textwrap.dedent(
-        """
-        from deephaven import new_table
-        from deephaven.column import string_col
-        import importlib.metadata
-
-        def __DH_MCP_make_pip_packages_table():
-            names = []
-            versions = []
-            for dist in importlib.metadata.distributions():
-                names.append(dist.metadata["Name"])
-                versions.append(dist.version)
-            return new_table([
-                string_col("Package", names),
-                string_col("Version", versions),
-            ])
-
-        _pip_packages_table = __DH_MCP_make_pip_packages_table()
-        """
-    )
-    _LOGGER.info("Running pip packages script in session...")
-    await asyncio.to_thread(session.run_script, script)
-    _LOGGER.info("Script executed successfully.")
-    arrow_table = await get_table(session, "_pip_packages_table")
-    _LOGGER.info("Table retrieved successfully.")
-    return arrow_table
-
-
-async def get_dh_versions(session: Session) -> tuple[str | None, str | None]:
-    """
-    Retrieve the Deephaven Core and Core+ versions installed in a given Deephaven session.
-    These versions are retrieved by running a script in the session that queries the installed pip packages.
-
-    Args:
-        session (Session): An active Deephaven session object.
-
-    Returns:
-        Tuple[Optional[str], Optional[str]]: A tuple containing:
-            - The version string for Deephaven Core, or None if not found (index 0).
-            - The version string for Deephaven Core+, or None if not found (index 1).
-
-    Raises:
-        RuntimeError: If the session is invalid, closed, or unable to execute scripts.
-        Exception: Any exception raised by get_pip_packages_table, table conversion, or data parsing (e.g., communication errors, unexpected data format).
-        These exceptions will propagate to the caller for handling.
-
-    Example:
-        >>> core_version, coreplus_version = await get_dh_versions(session)
-        >>> print(core_version)
-        '0.39.0'
-        >>> print(coreplus_version)
-        '0.39.0'
-    """
-    arrow_table = await get_pip_packages_table(session)
-    dh_core_version: str | None = None
-    dh_coreplus_version: str | None = None
-
-    if arrow_table is not None:
-        df = arrow_table.to_pandas()
-        raw_packages = df.to_dict(orient="records")
-        for pkg in raw_packages:
-            pkg_name = pkg.get("Package", "").lower()
-            version = pkg.get("Version", "")
-            if pkg_name == "deephaven-core" and dh_core_version is None:
-                dh_core_version = version
-            elif (
-                pkg_name == "deephaven_coreplus_worker" and dh_coreplus_version is None
-            ):
-                dh_coreplus_version = version
-            if dh_core_version and dh_coreplus_version:
-                break
-
-    return dh_core_version, dh_coreplus_version
